Disable config file watcher if a hot-reload server capability is set
Corresponds to https://reviews.llvm.org/D94222

I've tested this locally with clangd with/without the capability.
