+++ src/index.ts
@@ -1,8 +1,8 @@
 import { commands, ExtensionContext, services, State, workspace } from 'coc.nvim';
-import { basename } from 'path';
 import * as cmds from './cmds';
 import { Ctx } from './ctx';
 import { FileStatus } from './file_status';
+import { ReloadFeature } from './reload';
 import * as install from './install';
 
 export async function activate(context: ExtensionContext): Promise<void> {
@@ -24,18 +24,14 @@ export async function activate(context: ExtensionContext): Promise<void> {
   }
 
   try {
-    await ctx.startServer(clangdPath);
+    await ctx.startServer(clangdPath, new ReloadFeature(ctx, () => activate(context)));
   } catch (e) {
     return;
   }
 
   const fileStatus = new FileStatus();
-  const fileWatcher = workspace.createFileSystemWatcher('**/{compile_commands.json,compile_flags.txt}');
-  fileWatcher.onDidChange((e) => reload(e.fsPath, ctx, context));
-  fileWatcher.onDidCreate((e) => reload(e.fsPath, ctx, context));
 
   context.subscriptions.push(
-    fileWatcher,
     fileStatus,
 
     commands.registerCommand('clangd.switchSourceHeader', cmds.switchSourceHeader(ctx)),
@@ -59,24 +55,4 @@ export async function activate(context: ExtensionContext): Promise<void> {
   ctx.client?.onNotification('textDocument/clangd.fileStatus', (status) => {
     fileStatus.onFileUpdated(status);
   });
-}
-
-async function reload(url: string, ctx: Ctx, context: ExtensionContext) {
-  const notification = ctx.config.showDBChangedNotification;
-  if (notification) {
-    const msg = `${basename(url)} has changed, clangd is reloading...`;
-    workspace.showMessage(msg);
-  }
-
-  for (const sub of ctx.subscriptions) {
-    try {
-      sub.dispose();
-    } catch (e) {
-      console.error(e);
-    }
-  }
-
-  await activate(context);
-
-  if (notification) workspace.showMessage(`clangd has reloaded`);
 }
